+++ src/main/java/com/example/springbootproject/auth/dto/request/RechargePointsRequest.java
@@ -1,4 +1,4 @@
 package com.example.springbootproject.auth.dto.request;
 
-public record RechargePointsRequest(int chargePoint) {
+public record RechargePointsRequest(Long chargePoint) {
 }
